Add fuller data listings
Increases amount of data from test ten records.
Also changes the titles on the sheets to something more JS friendly.

And... adds in a link to departmental pages, because... I was trying to
figure out who the Traffic Commisioners for Great Britian were and being
able to see their organistaion page helped.